Add Plop to generate components.
Making new components is a bit of a hassle with all the files that need to be seeded. This introduces Plop to streamline things a lot more.

All you need to do now is run `npm run plop`, follow the prompts and you're blessed with a shiny new component ready for your dreams to come true in.