Some tweaks to the Card component.
`width: 100%` needed to make cards all the same width when they have different text while in a flex row/column.

Added a param for the onClick event to the Card. I needed to do a preventDefault on the onClick of the Card I'm working with, but TypeScript complains because the types don't match. ( `() => void` vs. `(e: MouseEvent<HTMLElement>) => void`);