Changing the way we load `debounce` to avoid loading all of lodash. (#168)
Doing `import { debounce } from "lodash";` seems to always load all of lodash regardless of if you're only using part of it or not. Even though TypeScript seems to complain about `import debounce from "lodash/debounce" it seems to be the right way to load it to only get the functions you need.

Executing all of `lodash` was causing problems on production because a side-effect of loading `lodash` is that it will put itself on window as `_`. But we're already using `underscore` in Jobber, so this then clobbers the existing underscore instance.

Mostly this was fine until we hit some code that executes a function that existins in `underscore`, but doesn't exist in lodash.