TECH-22 Removing `no-null` linter rule
There was a lot of discussion around the rule and we eventually decided to take it out. Because we're using tools and libraries that aren't as strict about `null` as this rule it means there's lots of cases where `null` is valid and necessary value to assign to a variable. An example of this is null fields in GraphQL. If a field is nullable and you want to test the null case your test data will necessarily have `null` in it. The rule meant that we were regularly adding exceptions or doing other type acrobatics to make the linter happy.