fix: InputNumber doesn't run custom validations
If you define a custom validation function on InputNumber like this,
```
    <InputNumber
      placeholder="Custom validation function"
      validations={{
        validate: validationHandler,
      }}
    />
```

The validation function won't run. This was because there's a validation function already defined internally for the component and it doesn't merge the user-defined function with the internal function.

I had to add a function to merge the validation functions together because if I just spread the `validate` property like this,
```
      validations={{
        ...props.validations,
        validate: {
          ...props.validations?.validate,
          getOverLimitMessage,
        },
```
it fails in the cases where there's no `validate` value defined.