feat(components): InputText Props for Autocomplete
The `Autocomplete` component utilizes `InputText` behind the scenes
to provide the primary input element that is rendered.

We've encountered a scenario where the following props need to be
passed down to the underlying InputText element:

* `suffix` - To style the input with an icon suffix
* `prefix` - To enable a clear action for the input
* `inputRef` - A reference to the input element in the DOM so that
               focus can be activated programatically.

This commit exposes the props listed above on `Autocomplete` and
forwards them to the `InputText` child. As the current `name` prop
is also forwarded to the input, I opted to remove the explicit
reference in favour of forwarding the prop via the spread operator.

This commit also changes the JS doc-style comment for the reason
why max-statements is disabled as it was rendering in our StoryBook
documentation. We update the lint exception to match our current
conventions, where we increase the number of statements allowed
inline.