fix(components-native): avoid InputText re-renders (#1652)
* fix(components-native): avoid InputText re-renders

The InputText native component automatically trims whitespace
from values.

Currently, the onChangeText callback is invoked when the input
is blurred even when the value does not change when whitespace
is trimmed. This can lead to additional renders, which cause
further state-related issues down the line (i.e. when displaying
contextual menus, etc.).

This commit adds a guard clause that avoids invoking the
`onChangeText` callback on blur when the trimmed value is identical
to the original input value (hence no change actually occurred).

* fix(components-native): InputCurrency Specs

The specs for InputCurrency currently rely on buggy blur behaviour
which cause some test to break due to the changes in commit d2b70d6.

The test cases are generally performing expectations against the
input formatting, therefore blur events aren't necessary.