fix(components): Handle Combobox scrollBottom edge case
The bottom gradient that hints there is more options in the list was never disappearing when you reached the bottom (for me at least). After logging out the scrollTop and clientHeight, I noticed it was 0.5px greater than the scrollHeight which failed this case. Checking for >= should handle these pixel rounding edge cases.