fix(components): Gracefully handle longer content in Combobox options [JOB-87773] (#1745)
* fix(components): Prevent Combobox checkmark from shrinking

When the text content for the selected Combobox option was a certain length, it would cause the checkmark to shrink because it's a flex row layout and the svg icon was adapting to fit.

Wrapping in a div prevents this.

* fix(components): Allow Combobox option height to grow with content

The ComboboxContentList ul specifies a max-height. Because of that, these options are using their own min-height and not growing when the content is larger as they should.

Using flex-shrink: 0 prevents the list options from shrinking to the min-height if the content is larger.

* fix(components): Handle Combobox scrollBottom edge case

The bottom gradient that hints there is more options in the list was never disappearing when you reached the bottom (for me at least). After logging out the scrollTop and clientHeight, I noticed it was 0.5px greater than the scrollHeight which failed this case. Checking for >= should handle these pixel rounding edge cases.