chore: Update actions to fix node 16 warnings (#1735)
* chore: Update actions to fix node 16 warnings

I believe we need to migrate to actions/checkout@v4 released Oct 2023
https://github.com/actions/checkout/releases/tag/v4.0.0

And same with actions/setup-node@v4 which internally uses actions/checkout@v4 now. This one is kind of odd since we do specify `with: node-version: 18` in the action... but maybe its internal dependency on checkout@v3 is causing the warning.

https://github.com/actions/setup-node/releases/tag/v4.0.0

* chore: Update codeql workflow to v3

To fix node 16 warnings.

* chore: Update actions/github-script to v7

To update the internal node version to v20.