Fix invalid @rollup/plugin-commonjs version
When running npm ls, a few dependencies are listed as "invalid".

This is because npm is checking against the package-lock, which contains the old version even though the design/package.json has ^28.0.0 listed.

I first deleted that library from design/package.json, ran npm install in the root to update the package-lock, and then added it back to design/package.json. This fixed the package-lock and removed the outdated version.