Fix typescript build issues
npm run build is currently broken.. while npm run dev works fine. This is because vite dev mode doesn't report typescript errors and doesn't try to make a production bundle.

`tsc -b` is the first thing that fails. This commit fixes those errors.

I had to upgrade to ts 5 for this package because downgrading vite to ts4 proved challenging and was throwing more errors. Upgrading this to ts5 actually isn't a problem because it doesn't affect the rest of our packages.