chore: Add CI jobs for new docs site & uncommitted files (#2094)
* Build new docs site

* Check git status for uncommitted files

* Fix typescript build issues

npm run build is currently broken.. while npm run dev works fine. This is because vite dev mode doesn't report typescript errors and doesn't try to make a production bundle.

`tsc -b` is the first thing that fails. This commit fixes those errors.

I had to upgrade to ts 5 for this package because downgrading vite to ts4 proved challenging and was throwing more errors. Upgrading this to ts5 actually isn't a problem because it doesn't affect the rest of our packages.

* Fix lint warnings

* Fix npm run build

The second failure was with `vite build` and was due to a missing alias for this formatters package.

* Report unused eslint directives

* Remove unnecessary step

We don't need this anymore. I think it was related to an old package we used to depend on.

* Improve type safety with satisfies