Fix toolbar's top gradient blend
In the previous commit, I forgot to test scrolling text below the toolbar and noticed the gradient blend wasn't working because the textarea height didn't include it.

This commit fixes the textarea's height and adds padding-bottom/scroll-padding-bottom to restore the behaviour where you're adding new lines to the bottom and chrome/safari will make sure the text is always visible as you're typing. Things brings back the firefox issue I mentioned in that earlier commit RE: scroll padding. However, it's not as bad as before because the padded area is only the height of the gradient blend... so the text is partially visible. But to make it less bad, I decided to cut that gradient in half for firefox only, and it looks not too bad now.