Place resize handle back on the textarea input
This actually reverts some changes made in an earlier PR: https://github.com/GetJobber/atlantis/pull/1856

In that PR, the resize handle was moved to the "textarea" class, but that's multiple levels above the actual textarea input element. This was the cause of the safari scrollbar bug, where the scrollbar overlapped the resize handle, blocking it from being used. However, as far as I can tell, that was done to avoid the firefox bug where the resize handle is sitting slightly above the toolbar... which is now back.

This change allowed me to delete all of the safari-specific css which removed the bad bottom padding which allows the text to scroll/flow seamlessly below the bottom of the element, exactly like chrome.

Given the minefield of browser compatibility issues around scroll-padding, scrollbar behaviour, resize behaviour, etc... I think firefox having the resize handle above the toolbar actually isn't terrible and is the least problematic of the alternatives. One browser-specific quirk is better than two.