Catch paths that don't resolve on the host
By copying index.html to 404.html, when the site is served at /docs and you try to load /docs/anything, it'll fallback to returning that 404 file which will render the SPA on that path.

This prevents the case where the host is falling back to the index.html at the root, which is Storybook, not our docs site.