Fix iframe code update
First, we were skipping the code update if the code was identical, which wasn't taking into account web and mobile examples being the same. We can now bypass that with the forceUpdate param.

Next, when the iframe's `load` event listener fires, we need to inject the iframe code.

Before, there was a race between the `load` event and the code below, which was firing a postMessage even though the frame wasn't ready for it yet.