Place resize handle back on the textarea input
This actually reverts some changes made in an earlier PR: #1856

In that PR, the resize handle was moved to the "textarea" class, but that's multiple levels above the actual textarea input element. This was the cause of the safari scrollbar bug, where the scrollbar overlapped the resize handle, blocking it from being used. However, as far as I can tell, that was done to avoid the issue where the resize handle is sitting slightly above the toolbar... which is now back.

But because we're moving towards a floating toolbar, the resize handle issue is no longer a problem, visually. And this change allowed me to delete all of the safari-specific css.