Reduce XOR complexity related to lack of jsdocs
Zak found that jsdoc comments for the title props weren't coming through his IDE and I confirmed mine as well. This appears to be a bug related to the typescript language server and complex XOR types.

We opted to reduce XOR complexity and just include the props on the main base type. Added react logic to prevent titleMetaData from being used with non-string titles.