chore: Fix test:visual command (#2629)
Fix test:visual command

It was failing because some of the build scripts require npm deps that are properly compiled for the docker container's linux environment.

This is actually something I had in place back in May (as part of Scott's original PR) but I reverted it because it started working without the npm install step.

I think something possibly recently changed regarding deps we have, OR I was just hitting a weird cache at the time which made me think it was unnecessary.