Ignore invalid values coming from useResizeObserver
I found this bug while investigating problems with our Banner screenshot tests. In some cases the icon appears in the wrong position but only in playwright's chromium screenshot.

Banner applies a class based on a width value that is returned from useResizeObserver. This class is constantly fluctuating the icon up and down each time the resize observer width changes. I added a bunch of logging to that hook and to the use-resize-observer dependency to figure out where the "1" width was coming from. I disabled the polyfill import we had and narrowed the problem down to the browser's ResizeObserver returning 1 unexpectedly. It's not clear why it's doing that, but we should be safe to exclude 1 and other invalid values.

Note: use-resize-observer is no longer maintained. Upgrading doesn't make sense, we should replace it with a modern version.