docs: Storybook v9 for Button (web) (#2676)
* skipping test

* Revert components-native changes

This PR is focusing on web's storybook first.

* Fix line endings

* Copy Button.stories.mdx to docs site

* Undo skipped text

* Upgrade storybook to latest v9 version

* Remove "web" subtree on new storybook

* Update old Button storybook docs to point to new site

* New storybook doesn't need to look for mdx files

Docs will only exist on the docs site, not in the new storybook instances.

* Activate a11y addon for storybook

* Uninstall storybook's onboarding addon

This is for learning how to use storybook, we don't need it.

* Disable storybook telemetry

* Remove docs addon

We aren't using docs in storybook v9. Our docs site handles that.

* Don't emit d.ts files for stories

* Cleanup imports

* Modernize story format

* import from "@storybook/react-vite"
* removed parameters
* removed decorators (unsure if we need that yet)
* adjusted the Story type
* used fn() for actions

* Update storybook config for components

* Disable special docs site storybook link handling for now

We'll do this as part of the storybook deployment ticket.

* Don't load broken component styles.css

Component styles.css are generated by rollup and have completely different classnames than what storybook is loading. Because storybook loads components from _src_, it processes their css in its own way and doesn't end up with the same classnames, so the dist/styles.css isn't actually doing anything.

We're very likely going to need to get vite compiling with postcss to match our component rollup config. Notably storybook v7 main.ts has webpack config for postcss, so we could potentially do the same thing with vite.

* Update controls config

* Restore ComponentLinks for now

We'll do this as part of the storybook deployment ticket.

* Cleanup

* Cleanup failing test

This test started failing because of updated sub dependencies as far as I can tell.

* Add theme toggle to storybook

* Remove unnecessary height

storyWrapper ensures full height

---------

Co-authored-by: Scott Thompson <145565743+scotttjob@users.noreply.github.com>