docs: Storybook v9 for Button (mobile) (#2682)
* skipping test

* Revert components-native changes

This PR is focusing on web's storybook first.

* Fix line endings

* Copy Button.stories.mdx to docs site

* Undo skipped text

* Upgrade storybook to latest v9 version

* Remove "web" subtree on new storybook

* Update old Button storybook docs to point to new site

* New storybook doesn't need to look for mdx files

Docs will only exist on the docs site, not in the new storybook instances.

* Activate a11y addon for storybook

* Uninstall storybook's onboarding addon

This is for learning how to use storybook, we don't need it.

* Disable storybook telemetry

* Remove docs addon

We aren't using docs in storybook v9. Our docs site handles that.

* Don't emit d.ts files for stories

* Cleanup imports

* Modernize story format

* import from "@storybook/react-vite"
* removed parameters
* removed decorators (unsure if we need that yet)
* adjusted the Story type
* used fn() for actions

* Update storybook config for components

* Disable special docs site storybook link handling for now

We'll do this as part of the storybook deployment ticket.

* Don't load broken component styles.css

Component styles.css are generated by rollup and have completely different classnames than what storybook is loading. Because storybook loads components from _src_, it processes their css in its own way and doesn't end up with the same classnames, so the dist/styles.css isn't actually doing anything.

We're very likely going to need to get vite compiling with postcss to match our component rollup config. Notably storybook v7 main.ts has webpack config for postcss, so we could potentially do the same thing with vite.

* Update controls config

* Restore ComponentLinks for now

We'll do this as part of the storybook deployment ticket.

* Cleanup

* Cleanup failing test

This test started failing because of updated sub dependencies as far as I can tell.

* Add theme toggle to storybook

* Install storybook for components-native

* Don't compile stories

* Don't publish new versions due to storybook changes

* Fix error with missing @storybook/react package

* Polyfill react-native for storybook

* Configure storybook and preview

* Remove unnecessary style

SBProvider supplies the background color via the themed token

* Shim react-native-web via vite plugin config

* Cleanup plugin handling

* Default to mobile viewport

* Add Button mobile story

* Cleanup plugin filter

* Activate @storybook/addon-a11y

* Type args

* Enable eslint formatting for components-native/storybook

* Get reanimated working

* Disable some eslint rules for .storybook files

* Configure eslint for web v9 .storybook directory

* Update package lock

* Upgrade react-native-web for storybook v7

Storybook v9 uses the v0.20 of RNW and it was leaking into the v7 SB webpack bundle due to module resolution - components-native's node_modules had react-native-web v0.20 which SB v7 loaded instead of the root node_modules version.

* Make docs site compatible with latest react-native-web

Same things I had to mock with v9 storybook

* Rename file

* Fix fn() actions by shimming react-native-reanimated internals

* Fix comment

---------

Co-authored-by: Scott Thompson <145565743+scotttjob@users.noreply.github.com>