Remove unnecessary extra wrapping element
InternalDynamicThemeProvider still needs a wrapping element because it doesn't have one at all in the normal case.

InternalStaticThemeProvider already has a div wrapper for the data-theme, so it makes sense to reuse that and just inject the css var overrides in that instead of another div.

Also reverted from span to div because I realized you can't render divs inside spans, so most use cases would be invalid.