feat(components): React 19 support (#2687)
* Upgrade packages to react 19

Notably the root package.json is still on react 18 for storybook v7 compatibility.

* Support both react 18 and 19

* Force packages to use react 19 in devDeps for testing

This ensures they use react 19 for tests, instead of falling back to react 18 in the root of the project.

* Fix docs site build error with react 19

This dependency is apparently supposed to be externalized and not included within the compiled bundles. Apparently for both react 18 and react 19.

It was causing errors in react 19.

* Fix missing createRoot function

createRoot moved to react-dom/client, so just had to fix this import.

Note this matches editorMobileBundle.

* Fix web and mobile iframe editors

They were loading multiple instances of react 19 from various directories I guess.

* Remove reference to old React 18 API

This code appears unnecessary as far as I can tell. Switching between web and mobile iframe editors now work under react 19.

* Npm install to clean up package-lock

* Move react 19 to root

* npm install to regen package-lock

* Delete and recreate package-lock

However I did UNDO deleted lines related to `@esbuild` deps. CI needs those for linux and there's a problem with npm not re-adding them as expected.

* Update react test renderer types

* Upgrade react-intl

* Upgrade testing library

* Remove unused storybook dep

* Move rollup dep to root

* Restore missing deps

Because I deleted the main package-lock.json and recreated it, I needed to restore some missing deps that npm mistakenly removed.

At the root: npm i rollup@4.24.0
This restores rollup binary deps.

In packages/site: npm i vite@5.3.1
This restores esbuild binary deps.

* Move framer-motion back to dep

To be consistent with the upgrade that is going to hit master soon.

Ideally we're not supposed to be dep+peerDep but rather devDep+peerDep. We're going to do a mass cleanup after react 19 lands.

* Make peerDep match

* Remove duplicate reference

* Fix lint warnings

* Remove bad vite version

Running `npm i vite@5.3.1` in packages/site caused some problems. That version was out of date and didn't match the version hoisted to root node_modules

* fix(components-native): Upgrade testing library and related dependencies for react 19 compatibility (#2752)

* Update react test renderer types

* Upgrade react-intl

* Upgrade testing library

* Remove unused storybook dep

* Fix jest-dom import

* Fix old @testing-library/react-hooks imports

Dep was deprecated and renderHook was added to testing-library/react

* Replace deprecated waitForNextUpdate with alternative solution

* Force override react-is to latest react 19 compatible version

* Update snapshots

* Fix react id format

They switched from :id: to this new format

* Fix renderHook imports for react-native

* Align react types at the root

* Upgrade react-native and react-native-gesture-handler

Matches our mobile app

* Upgrade @react-native/babel-preset

* Regen parts of package-lock

* Fix module path

* Downgrade react to 19.1.0 for compatibility with react-native-gesture-handler

Tests were failing with incompatible react versions.

* Update datetimepicker related stuff

* Update snapshot

* Update another snapshot

* Temp skip failing components-native test suites

* Fix missing jest-dom types

* Fix Thumbnail tests to wait for element

ThumbnailImage.tsx has a useEffect on mount that sets state. Had to wait for that to run for the test to be successful.

* Fix EmptyState tests

This was spying on Icon before which is not ideal. Instead, to confirm the color and size match what we expect, I dug down into those elements to check their values.

* Remove deprecated package @testing-library/jest-native

* Fix useScrollToError tests

Missing accessibilityFocus method

* Fix tests by mocking react-native-modalize

* Fix AutoLink tests

* Fix CheckboxGroup test

* Fix react-native Modal mock

* Fix Prefix tests

Replace broken spy

* Fix FormMessage tests

Needs to wait for the element to become visible

* Fix FormatFile tests

* we were expecting createThumbnail to be called twice for some reason?
* useCreateThumbnail is a dynamically generated hook which goes against the rules of hooks. This needs to be cleaned up, i'll log a note for it

* Align @react-native/babel-preset

* Fix typescript compilation errors

tsc was erroring due to me adding the types array which explicitly lists which types to use. Had to add node and bump es2018 to es2020 to make it understand certain parts of code.

* Fix components bootstrap

incompatible tslib version was at the root, causing rollup to fail

* Acknowledge typescript error and temp disable it to get compilation to succeed

* Fix some lint errors

* Update package-lock.json

Some packages were under `packages/components-native` and needed to be hoisted to the root node_modules to be resolved by other packages (like docs site, etc)

react-native, react-native-portalize, react-native-modalize

I had to temp install them at the root with `npm i <pkg> -D -f` to get the lockfile to hoist them there.

* Fix path now that modules were hoisted to root again

* Resolve missing module

Had to npm install storybook inside packages/components

* Regen props

* Hoist shared storybook deps to root

* Revert "Hoist shared storybook deps to root"

This reverts commit 32ed93c65b62e01bfb8bc293e9a926647f530eec.

Had to revert this for now. It's not necessary and causing peerDep conflicts in CI.

* Improve test query for clarity

* build: Fix cloudflare build failures with react 19 (#2756)

* Debug cloudflare failure

* Try removing root node_modules to see what's missing from deps

* Disable test

* Try adding missing package as explicit devDep

* Add missing package as explicit dep to fix CF pages build

* Restore build script

* Try to fix CF build error

CF seems to not correctly install dependencies when they are marked as peer?

* fix(components-native): Fix typescript error in components native (#2762)

* fix onFocus callback type

* update props json

---------

Co-authored-by: Edi Yang <edison.cy.yang@gmail.com>