Replace border with box-shadow
The border was causing an overflow x scrollbar to appear in certain conditions, notably when the browser is zoomed in at various levels (probably related to pixel rounding).

I didn't want to add `overflow: hidden` to the Card itself because that could introduce downstream breakages in consuming apps if they rely on that behaviour at all.

Instead, I solved this by replacing border with box-shadow which avoids causing overflow issues.