fix(components): Prevent Card accent from causing overflow under certain conditions (#2802)
* Add accent story

* Add box-sizing to card

We do this in our consuming app, so applying it here makes it more consistent with what we ship.

The visual test updates look correct to me. If you replace those Cards with Boxes, you'll see the gap lines up exactly with the Cards when box-sizing is set.

* Replace border with box-shadow

The border was causing an overflow x scrollbar to appear in certain conditions, notably when the browser is zoomed in at various levels (probably related to pixel rounding).

I didn't want to add `overflow: hidden` to the Card itself because that could introduce downstream breakages in consuming apps if they rely on that behaviour at all.

Instead, I solved this by replacing border with box-shadow which avoids causing overflow issues.