Fix shadow flicker
Very buggy problem with the scrollview. If the modal is at the top/fullscreen, and you drag it downwards then upwards, it rapidly fluctuates nativeEvent.contentOffset.y as if you're scrolling... but you're not.. the modal is just moving its position back to the top.

This solution doesn't fully cover it, so perhaps we need a ticket to track a followup improvement here.